Completed uniqueCongress API call.
Completed uniqueCongress API call and have commented all the neccesary portions. Note: mongoDB uri used is a local link and will need to be connected to the server uri. The collections data is the senate/house stock watcher "all_transactions" JSON file and have been labeled "houseStockWatcher" and "senateStockWatcher" in the database labeled "congressStockWatcher". Also Note: To test, you must run the client and server and travel to the "Politicians" tab in order for the uniqueCongress array of JSON data to be console logged.