Connected stockAct and propublica APIs. There are 9 names that did not match.
Connected stockAct and propublica APIs. There are 9 names that did not match. They did not match because the 9 names are politicians from older congress like the 115th or 116th congress. May plan on fixing it by making more API calls to propublica for those versions of congress. Also, still need to comment.