Completed and cleaned up mapping file in mongodb. Completed uniqueCongress call. Removed mapping.js file elsewhere.
The mapping file in the database is completed and cleaned up where it is 99% accurate information and now has politician portrait uri's attached the the unique politician objects. The virtualDatabase.js file is able to do the uniqueCongress call and send back the specific object back to the client. Removed unnessary code in virtualDatabase.js file now that the mapping file in mongodb works. Removed mapping.js file. The mapping.js file is an individual file that was supposed to be used to push the mapping file to the mongodb but was not because it currently does not work. Decided to move the mapping.js file elsewhere because it is not needed for any other functionality. Will continue to try getting it to work because it may be important in explaining how our mapping file was created.